[f] Move build diffs logic to Equal
Only Equal needs to build diffs, so let it do that by itself.
