+++ robber/constants.py
@@ -1,2 +1,2 @@
 TENSE_WORDS = ['be', 'have', 'been']
-DIFF_BUILDABLE_TYPES = [str, list, dict]
+DIFF_BUILDABLE_TYPES = [str, list]
